+++ robots/src/gui/model/BacteriaCellEntity.java
@@ -1,7 +1,12 @@
 package gui.model;
 
 public class BacteriaCellEntity extends CellEntity {
-    public BacteriaCellEntity() {
+    /*public BacteriaCellEntity() {
         super(CellType.BACTERIA);
+    }*/
+
+    @Override
+    public void update(WorldContext context) {
+
     }
 }
