--- robots/src/gui/model/EmptyCellEntity.java
@@ -1,7 +1,0 @@
-package gui.model;
-
-public class EmptyCellEntity extends CellEntity {
-    public EmptyCellEntity() {
-        super(CellType.EMPTY);
-    }
-}
