+++ robots/src/gui/model/FoodCellEntity.java
@@ -1,7 +1,12 @@
 package gui.model;
 
 public class FoodCellEntity extends CellEntity{
-    public FoodCellEntity() {
+    /*public FoodCellEntity() {
         super(CellType.FOOD);
+    }*/
+
+    @Override
+    public void update(WorldContext context) {
+
     }
 }
