--- robots/src/gui/view/WorldRenderer.java
@@ -1,12 +1,0 @@
-package gui.view;
-
-import gui.model.WorldEntity;
-
-import java.awt.*;
-
-public class WorldRenderer implements EntityRenderer<WorldEntity> {
-    @Override
-    public void render(WorldEntity entity, Graphics graphics) {
-
-    }
-}
